Grosse update en lien avec les modifs en cours de Michael (->doubles): Passage de tous les QVector2D ou QVector3D en Eigen::Vector2d ou Eigen::Vector3d -> Tous les items, drawmanagers, fonctions ct_math/* (en en profitant pour y virer les templates)
git-svn-id: http://rdinnovation.onf.fr/svn/computree@253 0d0f61bb-bbfd-4940-b5ed-fa75979e1372

+++ pluginshared/ct_itemdrawable/abstract/ct_abstractareashape2d.cpp
@@ -58,7 +58,7 @@ QString CT_AbstractAreaShape2D::staticGetType()
     return CT_AbstractItemDrawableWithoutPointCloud::staticGetType() + "/CT_AbstractAreaShape2D";
 }
 
-bool CT_AbstractAreaShape2D::contains(float x, float y) const
+bool CT_AbstractAreaShape2D::contains(double x, double y) const
 {
     return getAreaData().contains(x, y);
 }
